Try to rename raster.py in raster to avoid errors on Mac and Windows
