+++ geoutils/raster/__init__.py
@@ -1,4 +1,4 @@
-from geoutils.raster.raster import Raster, RasterType, Mask  # noqa isort:skip
+from geoutils.raster.core import Raster, RasterType, Mask  # noqa isort:skip
 from geoutils.raster.array import *  # noqa
 from geoutils.raster.multiraster import *  # noqa
 from geoutils.raster.sampling import *  # noqa
